backend/requirements.txt
fastapi
uvicorn[standard]

SQLAlchemy
psycopg[binary]

python-dotenv
pydantic

alembic

passlib[argon2]

pytest
httpx
